fix: Moved membership code to separate files
