+++ terraso_backend/apps/graphql/schema/groups.py
@@ -15,31 +15,16 @@
 
 import django_filters
 import graphene
-import rules
 import structlog
-from django.core.exceptions import NON_FIELD_ERRORS, ValidationError
-from django.db import IntegrityError, transaction
+from django.db import transaction
 from graphene import relay
 from graphene_django import DjangoObjectType
 
-from apps.collaboration.graphql import CollaborationMembershipNode
-from apps.collaboration.models import Membership as CollaborationMembership
 from apps.collaboration.models import MembershipList
-from apps.core import group_collaboration_roles
 from apps.core.models import Group
-from apps.graphql.exceptions import (
-    GraphQLNotAllowedException,
-    GraphQLNotFoundException,
-    GraphQLValidationException,
-)
-from apps.notifications.email import EmailNotification
+from apps.graphql.exceptions import GraphQLNotAllowedException
 
-from .commons import (
-    BaseAuthenticatedMutation,
-    BaseDeleteMutation,
-    BaseWriteMutation,
-    TerrasoConnection,
-)
+from .commons import BaseDeleteMutation, BaseWriteMutation, TerrasoConnection
 from .constants import MutationTypes
 from .shared_resources_mixin import SharedResourcesMixin
 
@@ -205,208 +190,3 @@ class GroupDeleteMutation(BaseDeleteMutation):
             )
 
         return super().mutate_and_get_payload(root, info, **kwargs)
-
-
-class GroupMembershipSaveMutation(BaseAuthenticatedMutation):
-    model_class = CollaborationMembership
-    memberships = graphene.Field(graphene.List(CollaborationMembershipNode))
-    group = graphene.Field(GroupNode)
-
-    class Input:
-        user_role = graphene.String()
-        user_emails = graphene.List(graphene.String, required=True)
-        group_slug = graphene.String(required=True)
-        membership_status = graphene.String()
-
-    @classmethod
-    @transaction.atomic
-    def mutate_and_get_payload(cls, root, info, **kwargs):
-        user = info.context.user
-
-        user_role = (
-            kwargs["user_role"] if "user_role" in kwargs else group_collaboration_roles.ROLE_MEMBER
-        )
-
-        if user_role not in group_collaboration_roles.ALL_ROLES:
-            logger.info(
-                "Attempt to save Group Memberships, but user role is not valid",
-                extra={
-                    "user_role": user_role,
-                },
-            )
-            raise GraphQLNotAllowedException(
-                model_name=CollaborationMembership.__name__, operation=MutationTypes.UPDATE
-            )
-
-        group_slug = kwargs["group_slug"]
-
-        try:
-            group = Group.objects.get(slug=group_slug)
-        except Exception as error:
-            logger.error(
-                "Attempt to save Story Map Memberships, but story map was not found",
-                extra={
-                    "group_slug": group_slug,
-                    "error": error,
-                },
-            )
-            raise GraphQLNotFoundException(model_name=Group.__name__)
-
-        def validate(context):
-            if not rules.test_rule(
-                "allowed_to_change_group_membership",
-                user,
-                {
-                    "group": group,
-                    **context,
-                },
-            ):
-                raise ValidationError("User cannot request membership")
-
-        is_closed_group = (
-            group.membership_list.membership_type == MembershipList.MEMBERSHIP_TYPE_CLOSED
-        )
-
-        membership_status = (
-            CollaborationMembership.APPROVED
-            if not is_closed_group
-            else kwargs["membership_status"]
-            if "membership_status" in kwargs
-            else CollaborationMembership.PENDING
-        )
-
-        try:
-            memberships_save_result = [
-                {
-                    "membership": result[1],
-                    "context": result[0],
-                }
-                for email in kwargs["user_emails"]
-                for result in [
-                    group.membership_list.save_membership(
-                        user_email=email,
-                        user_role=user_role,
-                        membership_status=membership_status,
-                        validation_func=validate,
-                    )
-                ]
-            ]
-        except ValidationError as error:
-            logger.error(
-                "Attempt to save Group Memberships, but user is not allowed",
-                extra={"error": str(error)},
-            )
-            raise GraphQLNotAllowedException(
-                model_name=CollaborationMembership.__name__, operation=MutationTypes.UPDATE
-            )
-        except IntegrityError as exc:
-            logger.info(
-                "Attempt to save Group Memberships, but it's not unique",
-                extra={"model": CollaborationMembership.__name__, "integrity_error": exc},
-            )
-
-            validation_error = ValidationError(
-                message={
-                    NON_FIELD_ERRORS: ValidationError(
-                        message=f"This {CollaborationMembership.__name__} already exists",
-                        code="unique",
-                    )
-                },
-            )
-            raise GraphQLValidationException.from_validation_error(
-                validation_error, model_name=CollaborationMembership.__name__
-            )
-        except Exception as error:
-            logger.error(
-                "Attempt to update Story Map Memberships, but there was an error",
-                extra={"error": str(error)},
-            )
-            raise GraphQLNotFoundException(model_name=CollaborationMembership.__name__)
-
-        if group.membership_list.membership_type == MembershipList.MEMBERSHIP_TYPE_CLOSED:
-            for membership_result in memberships_save_result:
-                context = membership_result["context"]
-                membership = membership_result["membership"]
-                if context["is_new"]:
-                    EmailNotification.send_membership_request(membership.user, group)
-                if context["is_membership_approved"]:
-                    EmailNotification.send_membership_approval(membership.user, group)
-
-        return cls(
-            memberships=[
-                membership_result["membership"] for membership_result in memberships_save_result
-            ],
-            group=group,
-        )
-
-
-class GroupMembershipDeleteMutation(BaseDeleteMutation):
-    membership = graphene.Field(CollaborationMembershipNode)
-    group = graphene.Field(GroupNode)
-
-    model_class = CollaborationMembership
-
-    class Input:
-        id = graphene.ID(required=True)
-        group_slug = graphene.String(required=True)
-
-    @classmethod
-    def mutate_and_get_payload(cls, root, info, **kwargs):
-        user = info.context.user
-        membership_id = kwargs["id"]
-        group_slug = kwargs["group_slug"]
-
-        try:
-            group = Group.objects.get(slug=group_slug)
-        except Group.DoesNotExist:
-            logger.error(
-                "Attempt to delete Group Membership, but group was not found",
-                extra={"group_slug": group_slug},
-            )
-            raise GraphQLNotFoundException(model_name=Group.__name__)
-
-        try:
-            membership = group.membership_list.memberships.get(id=membership_id)
-        except CollaborationMembership.DoesNotExist:
-            logger.error(
-                "Attempt to delete Group Membership, but membership was not found",
-                extra={"membership_id": membership_id},
-            )
-            raise GraphQLNotFoundException(model_name=CollaborationMembership.__name__)
-
-        if not rules.test_rule(
-            "allowed_to_delete_group_membership",
-            user,
-            {
-                "group": group,
-                "membership": membership,
-            },
-        ):
-            logger.info(
-                "Attempt to delete Group Memberships, but user lacks permission",
-                extra={"user_id": user.pk, "membership_id": membership_id},
-            )
-            raise GraphQLNotAllowedException(
-                model_name=CollaborationMembership.__name__, operation=MutationTypes.DELETE
-            )
-
-        if not rules.test_rule(
-            "allowed_group_managers_count",
-            user,
-            {
-                "group": group,
-                "membership": membership,
-            },
-        ):
-            logger.info(
-                "Attempt to update a Membership, but cannot remove last manager",
-                extra={"user_id": user.pk, "membership_id": membership_id},
-            )
-            raise GraphQLNotAllowedException(
-                model_name=CollaborationMembership.__name__,
-                operation=MutationTypes.DELETE,
-                message="manager_count",
-            )
-
-        result = super().mutate_and_get_payload(root, info, **kwargs)
-        return cls(membership=result.membership, group=group)
